requirements.txt
aws-cdk-lib>=2.170.0
constructs>=10.0.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
requests>=2.31.0
pyyaml>=6.0
mcp>=1.0.0
